refractory, more can see zcong1993.md
