+++ README.md
@@ -1,6 +1,18 @@
 # First-Nodejs-Expressjs
+
 This is my first server code with nodejs and expressjs.
 I decide to start to learn node!
 Thank you if you can provide any suggestion!
 If you are a beginner we can learn together.
 
+## start
+
+```bash
+# clone first
+$ git clone https://github.com/chinazhangyujia/First-Nodejs-Expressjs.git
+$ cd First-Nodejs-Expressjs
+# install dependencies
+$ npm install
+# serve
+$ npm start # or node app.js
+```
