+++ Resources/email.json
@@ -1,11 +1,14 @@
 {
-  "Email1" : {
-    "account": "yujia5@ufl.edu",
-    "password": "Rvy-Vty++x"
-  },
-  "Email2": {
-    "account": "chinazhangyujia@gmail.com",
-    "password": "Zyj931004"
-  }
+    "Email1": {
+        "account": "yujia5@ufl.edu",
+        "password": "cdcdxs"
+    },
+    "Email2": {
+        "account": "chinazhangyujia@gmail.com",
+        "password": "cdxs"
+    },
+    "Email3": {
+        "account": "xs",
+        "password": "cdcd"
+    }
 }
-{"Email3":{"account":"2411537283@qq.com","password":"zyj3202"}}{"Email3":{"account":"2411537283@qq.com","password":"zyj3202"}}{"Email3":{"account":"","password":""}}
