--- View/index.html
@@ -1,14 +1,0 @@
-<html>
-<head>
-  <h1>Create Email</h1>
-</head>
-<body>
-
-<form name="myForm" action="/createAccount"  method="post">
-  account: <input type="text" name="account">
-  password: <input type = "text" name = "password">
-  <input type="submit" value="Submit">
-</form>
-
-</body>
-</html>
