+++ app.js
@@ -1,15 +1,23 @@
-var http = require('http');
 var express = require('express');
-var router1 = require('./Router/router1');
+var bodyParser = require('body-parser');
+var urlencodedParser = bodyParser.urlencoded({extended: false});
+var customRouter = require('./router');
 var PORT = 8000;
 var app = express();
 
+// use URL-encoded parser as a top-level middleware, which will parse the bodies of all incoming requests.
+app.use(urlencodedParser);
 
-app.use('/', router1);
+// serve static file such as image, now you can use img in views by the url '/static/Admirals.PNG'
+app.use('/static', express.static('static'));
+
+app.use('/', customRouter);
 app.get('/', function(req, res) {
-  res.sendfile('./View/email.html', {root: __dirname});
-})
+  // views folder is recommended for html view file
+  res.sendFile('./index.html', { root: __dirname + '/views/' });
+});
 
-http.createServer(app).listen(PORT, function(){
+// experss can create http server itself
+app.listen(PORT, function(){
   console.log('listening port '+ PORT);
 });
