+++ package.json
@@ -1,0 +1,12 @@
+{
+  "name": "First-Nodejs-Expressjs",
+  "description": "This is my first server code with nodejs and expressjs. I decide to start to learn node! Thank you if you can provide any suggestion! If you are a beginner we can learn together.",
+  "main": "app.js",
+  "scripts": {
+    "start": "node app.js"
+  },
+  "dependencies": {
+    "body-parser": "^1.17.1",
+    "express": "^4.15.2"
+  }
+}
