+++ views/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>first-nodejs-express</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+  </head>
+  <body>
+    <h1>Create Email</h1>
+    <img src="/static/Admirals.PNG" alt="">
+
+    <form name="myForm" action="/createAccount"  method="post">
+      account: <input type="text" name="account">
+      <!--password should use type password-->
+      password: <input type = "password" name = "password">
+      <input type="submit" value="Submit">
+    </form>
+  </body>
+</html>
