+++ zcong1993.md
@@ -1,0 +1,19 @@
+##  edited by zcong1993
+
+> just did the following things make the repos more clean and friendly for beginers
+
+### 1. add `package.json`
+
+friendly for others, cause people can install all 3rd part packages your project used only use `npm install`
+
+### 2. refractory project folder and file
+
+use names which are more readable or offical recommended
+
+### 3. refractory json data store
+
+see `router/index.js`
+
+### 4. more
+
+see the comments
